+++ src/com/ihunda/android/binauralbeat/viz/HypnoticSpiral.java
@@ -84,7 +84,7 @@ public class HypnoticSpiral implements Visualization {
 	 * Beat frequency in Hz
 	 */
 	float freq;
-	double period;
+	float period;
 	
 	Bitmap preCalc[];
 	
@@ -101,22 +101,22 @@ public class HypnoticSpiral implements Visualization {
 	}
 	
 	private Bitmap getBitmap(int c, int width, int height) {
-		double w2=width/2;
-	    double h2=height/2;
-	    double w = 20; //(mouseX/5.0);
-	    double wPI = w/Math.PI;
+		float w2=width/2;
+	    float h2=height/2;
+	    float w = 20; //(mouseX/5.0);
+	    float wPI = (float) (w/Math.PI);
 	    
 	    int img[] = new int[width*height];
 		    		    
 	    int base = 0;
 	    for (int y = 0; y < height; y++) {
-	    	double dy = h2-y;
+	    	float dy = h2-y;
 
 	    	for (int x = 0; x < width; x++) {
-	    		double dx = w2-x;
+	    		float dx = w2-x;
 
-	    		double angle = c + Math.atan2(dy,dx) * wPI;
-	    		double dist = Math.sqrt(dx*dx+dy*dy);
+	    		float angle = c + (float) (Math.atan2(dy,dx)) * wPI;
+	    		float dist = (float) (Math.sqrt(dx*dx+dy*dy));
 	    		int d = (int) ((angle+dist)/w);
 	    		if (( d & 1 ) == 1) {
 	    			img[x+base] = Color.WHITE;
@@ -132,8 +132,8 @@ public class HypnoticSpiral implements Visualization {
 	
 	public void redraw(Canvas ca, int width, int height, float now,
 			float totalTime) {
-		double dperiod = period * 60;
-		double ratio = (now % dperiod) / dperiod;
+		float dperiod = period * 60;
+		float ratio = (now % dperiod) / dperiod;
 		
 		int c = (int) (ratio*NUM_PRECALC_BITMAPS);
 		Bitmap bm = preCalc[c];
